Finished basic registration flow for buyers
The basic UI is there and fully functional for
the buyers to create an account and list their
vehicle data which I then store in our database
(Firestore).

We still need to work on the validation for all
inputs and have better inputs instead of plain
text but I guess having a functional app is more
important than all these little parts.